requirements.txt
Flask==1.0.2
boto3
pillow